Resolve "Add userAuth to certain endpoints, so that user can only access his/her info."

Closes #29

See merge request maraba/lyhike-nimi-back!39

+++ src/main/java/ee/veebiprojekt/veebiprojekttest/service/RatingService.java
@@ -2,10 +2,13 @@ package ee.veebiprojekt.veebiprojekttest.service;
 
 import ee.veebiprojekt.veebiprojekttest.dto.RatingDTO;
 import ee.veebiprojekt.veebiprojekttest.entity.Rating;
+import ee.veebiprojekt.veebiprojekttest.entity.User;
 import ee.veebiprojekt.veebiprojekttest.exception.EntityNotFoundException;
 import ee.veebiprojekt.veebiprojekttest.exception.InvalidValueException;
+import ee.veebiprojekt.veebiprojekttest.exception.WrongOwnerException;
 import ee.veebiprojekt.veebiprojekttest.mapper.RatingMapper;
 import ee.veebiprojekt.veebiprojekttest.repository.RatingRepository;
+import ee.veebiprojekt.veebiprojekttest.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
@@ -17,12 +20,14 @@ public class RatingService {
     private final RatingRepository ratingRepository;
     private final RatingMapper ratingMapper;
     private final JokeService jokeService;
+    private final UserRepository userRepository;
     private static final String RATING = "Rating";
 
-    public RatingService(RatingRepository ratingRepository, RatingMapper ratingMapper, JokeService jokeService) {
+    public RatingService(RatingRepository ratingRepository, RatingMapper ratingMapper, JokeService jokeService, UserRepository userRepository) {
         this.ratingRepository = ratingRepository;
         this.ratingMapper = ratingMapper;
         this.jokeService = jokeService;
+        this.userRepository = userRepository;
     }
 
     private Rating authRatingDto(RatingDTO ratingDTO) {
@@ -38,17 +43,18 @@ public class RatingService {
         Rating rating = new Rating();
         rating.setRatingValue(ratingDTO.ratingValue());
         rating.setJokeId(ratingDTO.jokeId());
-        rating.setAuthorId(ratingDTO.authorId());
         log.debug("Authenticated rating: {}", rating);
         return rating;
     }
 
-    public RatingDTO addRating(RatingDTO ratingDTO) {
-        log.debug("Adding rating: {}", ratingDTO);
+    public RatingDTO addRating(RatingDTO ratingDTO, String username) {
+        log.debug("Adding rating: {} by {}", ratingDTO, username);
+        User user = userRepository.findByUsername(username);
         Rating rating = authRatingDto(ratingDTO);
+        rating.setAuthorId(user.getUserId());
         ratingRepository.save(rating);
         jokeService.recalculateJokeRating(ratingDTO.jokeId());
-        log.debug("Added rating: {}", rating);
+        log.debug("Added rating: {} by {}", rating, username);
         return ratingDTO;
     }
 
@@ -66,9 +72,11 @@ public class RatingService {
         return ratingRepository.findAll();
     }
 
-    public RatingDTO editRating(RatingDTO ratingDTO, long id) {
+    public RatingDTO editRating(RatingDTO ratingDTO, long id, String username) {
         log.debug("Editing rating with id {}", id);
         Rating rating = ratingRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(RATING, id));
+        User user = userRepository.findByUsername(username);
+        verifyRatingOwner(rating, user);
         if (ratingDTO.ratingValue() != null) rating.setRatingValue(ratingDTO.ratingValue());
         if (ratingDTO.jokeId() != null) rating.setJokeId(ratingDTO.jokeId());
         ratingRepository.save(rating);
@@ -77,12 +85,21 @@ public class RatingService {
         return ratingMapper.toDTO(rating);
     }
 
-    public RatingDTO deleteRating(Long id) {
+    public RatingDTO deleteRating(Long id, String username) {
         log.debug("Deleting rating with id {}", id);
         Rating rating = ratingRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(RATING, id));
+        User user = userRepository.findByUsername(username);
+        verifyRatingOwner(rating, user);
         ratingRepository.deleteById(id);
         jokeService.recalculateJokeRating(rating.getJokeId());
         log.debug("Deleted rating: {}", rating);
         return ratingMapper.toDTO(rating);
+    }
+
+    private void verifyRatingOwner(Rating rating, User user) {
+        if (rating.getAuthorId() != user.getUserId()) {
+            log.debug("User {} is not the author of rating with id {}", user.getFullName(), rating.getId());
+            throw new WrongOwnerException(RATING, rating.getId(), user.getUserId());
+        }
     }
 }
